% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{build_casecount_display}
\alias{build_casecount_display}
\title{Add a case count display using 'geo cards' to an app}
\usage{
build_casecount_display(
  app,
  sources,
  name,
  desc,
  ref_source = NULL,
  nrow = 2,
  ncol = 3,
  thumb = FALSE,
  state = NULL,
  views = NULL,
  id = NULL,
  order = 1,
  case_fatility_lim = 20
)
}
\description{
Add a case count display using 'geo cards' to an app
}
