% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{default_views}
\alias{default_views}
\title{Specify default views for a display}
\usage{
default_views(
  ref_source,
  comp_sources = NULL,
  death_cutoff = 100,
  entity_pl = "entities"
)
}
\arguments{
\item{ref_source}{A reference source against which to specify default views.}

\item{comp_sources}{Source IDs to compare the reference source to.}

\item{death_cutoff}{When sorting by case fatality rate, what cutoff of deaths to use for excluding low-count regions.}

\item{entity_pl}{The plural name to use for the geographic entity as it will be referred to in labels (e.g. "states", "countries", etc.)}
}
\description{
Specify default views for a display
}
