% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_scp}
\alias{deploy_scp}
\title{Copy app to a remote location over SCP}
\usage{
deploy_scp(app, ip, scp_token = NULL)
}
\arguments{
\item{app}{An app object created with \code{\link[=register_app]{register_app()}}.}

\item{ip}{IP address to scp to.}

\item{scp_token}{Token.}
}
\description{
Copy app to a remote location over SCP
}
