% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_links.R
\name{geo_link_href}
\alias{geo_link_href}
\title{Specify a cognostic that links to another display of a higher level of geographic resolution}
\usage{
geo_link_href(display, ref_level, desc = NULL, type = "href")
}
\arguments{
\item{display}{The name of the display to link to.}

\item{ref_level}{String that will be used in constructing the cognostic name (e.g. if ref_level = "who_regions", the cognostic will be called "view_who_regions").}

\item{desc}{An optional description for the cognostic.}

\item{type}{One of either "href" or "href_hash". "href" will open the link in a new page. "href_hash" will update the page's hash and reload the page (useful when changing state inside an iframe).}
}
\description{
Specify a cognostic that links to another display of a higher level of geographic resolution
}
