% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{register_app}
\alias{register_app}
\title{Register an app}
\usage{
register_app(name, path = tempfile())
}
\arguments{
\item{name}{App name.}

\item{path}{Directory in which app will be deployed.
By default, this is a temporary directory, and then the app can be
deployed to possibly several locations using \code{\link[=deploy_app]{deploy_app()}}.}
}
\description{
Register an app
}
