% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{source_entry}
\alias{source_entry}
\title{Specify a source entry}
\usage{
source_entry(source_id, admin_level, file)
}
\arguments{
\item{source_id}{A unique ID for the source.}

\item{admin_level}{One of "0", "1", "2", "global", "continent", "who_region".}

\item{file}{A URL or local file path to the data file for this source.}
}
\description{
Specify a source entry
}
