% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{source_list}
\alias{source_list}
\title{Specify a source list}
\usage{
source_list(...)
}
\arguments{
\item{\ldots}{objects created from \code{\link[=source_entry]{source_entry()}}}
}
\description{
Specify a source list
}
