% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{view_app}
\alias{view_app}
\title{Copies app to a location}
\usage{
view_app(app)
}
\arguments{
\item{app}{An app object created with \code{\link[=register_app]{register_app()}}.}
}
\description{
Copies app to a location
}
