% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_casecount_display}
\alias{build_casecount_display}
\title{Add a case count display using 'geo cards' to an app}
\usage{
build_casecount_display(
  app,
  sources,
  name,
  desc,
  ref_source = NULL,
  append_higher_admin_name = FALSE,
  max_date = NULL,
  min_date = NULL,
  geo_links = NULL,
  nrow = 2,
  ncol = 3,
  thumb = FALSE,
  state = NULL,
  views = NULL,
  id = NULL,
  order = 1,
  case_fatality_max = 20,
  require_token = FALSE,
  md_desc = ""
)
}
\arguments{
\item{app}{An app object created with \code{\link[=register_app]{register_app()}}.}

\item{sources}{A list of sources created with \code{\link[=source_list]{source_list()}}.}

\item{name}{A name for the display, which is displayed in the header of the app when this display is selected.}

\item{desc}{A description that will be shown under the name.}

\item{ref_source}{The name of the reference source, which should match one of the source_ids provided in \code{\link[=source_entry]{source_entry()}} when building the \code{\link[=source_list]{source_list()}}.}

\item{append_higher_admin_name}{If \code{TRUE}, will append the higher-level admin name to the geocard headers. For example, if currently creating a display for admin level 2 (e.g. US counties), setting to TRUE will append the state name to the geocard header. So if viewing for Benton County in the state of Washington, the header, instead of being simply being "Benton", will now be "Benton, Washington".}

\item{max_date}{Optional maximum date to plot. If not specified, will be set to latest date.}

\item{min_date}{Optional minimum date to plot. If not specified, will be set to 8 weeks prior to max date.}

\item{geo_links}{Optional object or list of objects created by \code{\link[=geo_link_href]{geo_link_href()}} or \code{\link[=geo_link_filter]{geo_link_filter()}} that will provide links to other displays (e.g. for a given continent, provide a link in its card that when clicked will open up the display for all countries in that continent).}

\item{nrow}{Default number of rows to show in the display.}

\item{ncol}{Default number of columns to show in the display.}

\item{thumb}{Optional path to an image file (png or jpeg) to use as the thumbnail for the display.}

\item{state}{Optional initial state (see the \code{state} argument of \link[trelliscopejs:trelliscope]{trelliscopejs::trelliscope} for more details).}

\item{views}{Optional pre-specified views to allow the user to choose between. See \code{\link[=default_views]{default_views()}} for more information.}

\item{id}{Optional ID passed to \link[trelliscopejs:trelliscope]{trelliscopejs::trelliscope}.}

\item{order}{Integer indicating what order this display should appear in in the display list shown when the app opens.}

\item{case_fatality_max}{The maximum y-axis limit to allow for case fatality percentage. Passed to \code{\link[geocard:geocard]{geocard::geocard()}}.}

\item{require_token}{Should a token be required to view the app? Passed to \link[trelliscopejs:trelliscope]{trelliscopejs::trelliscope}.}

\item{md_desc}{Optional longer-form description of the display, provided as a markdown string.}
}
\description{
Add a case count display using 'geo cards' to an app
}
