% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_casecount_display}
\alias{build_casecount_display}
\title{Add a case count display using 'geo cards' to an app}
\usage{
build_casecount_display(
  app,
  sources,
  name,
  desc,
  ref_source = NULL,
  append_higher_admin_name = FALSE,
  max_date = NULL,
  min_date = NULL,
  geo_links = NULL,
  nrow = 2,
  ncol = 3,
  disclaimer = FALSE,
  thumb = FALSE,
  state = NULL,
  views = NULL,
  id = NULL,
  order = 1,
  case_fatality_max = 20,
  md_desc = ""
)
}
\arguments{
\item{app}{An app object created with \code{\link[=register_app]{register_app()}}.}

\item{sources}{TODO}

\item{name}{TODO}

\item{desc}{TODO}

\item{ref_source}{TODO}

\item{append_higher_admin_name}{TODO}

\item{max_date}{TODO}

\item{min_date}{TODO}

\item{geo_links}{TODO}

\item{nrow}{TODO}

\item{ncol}{TODO}

\item{disclaimer}{TODO}

\item{thumb}{TODO}

\item{state}{TODO}

\item{views}{TODO}

\item{id}{TODO}

\item{order}{TODO}

\item{case_fatality_max}{TODO}

\item{md_desc}{TODO}
}
\description{
Add a case count display using 'geo cards' to an app
}
