% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{default_views}
\alias{default_views}
\title{Specify default views for a display}
\usage{
default_views(
  ref_source,
  comp_sources = NULL,
  death_cutoff = 100,
  entity_pl = "entities"
)
}
\arguments{
\item{ref_source}{TODO}

\item{comp_sources}{TODO}

\item{death_cutoff}{TODO}

\item{entity_pl}{TODO}
}
\description{
Specify default views for a display
}
