% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_cp}
\alias{deploy_cp}
\title{Copy app to a local location}
\usage{
deploy_cp(app, dest_dir, require_app_token = FALSE)
}
\arguments{
\item{app}{An app object created with \code{\link[=register_app]{register_app()}}.}

\item{dest_dir}{TODO}

\item{require_app_token}{TODO}
}
\description{
Copy app to a local location
}
