% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_netlify}
\alias{deploy_netlify}
\title{Deploy an app to Netlify}
\usage{
deploy_netlify(app, netlify_app_id, require_app_token = FALSE)
}
\arguments{
\item{app}{An app object created with \code{\link[=register_app]{register_app()}}.}

\item{netlify_app_id}{Netlify app ID.}

\item{require_app_token}{Netlify app token.}
}
\description{
Deploy an app to Netlify
}
